+++ tests/test_information.py
@@ -4,8 +4,8 @@ from ingenialink.ethercat.network import EthercatNetwork
 from ingenialink.ethernet.network import EthernetNetwork
 from ingenialink.canopen.network import CanopenNetwork, CAN_DEVICE, CAN_BAUDRATE
 from ingenialink.dictionary import SubnodeType
+from ingenialink.register import REG_ACCESS, REG_DTYPE
 
-from ingeniamotion.enums import REG_DTYPE, REG_ACCESS
 from ingeniamotion.information import COMMUNICATION_TYPE
 from ingeniamotion.exceptions import IMRegisterNotExist, IMException
 
@@ -243,7 +243,7 @@ def test_get_slave_id_exception(motion_controller):
 
 @pytest.mark.virtual
 def test_get_baudrate_success(motion_controller, mocker):
-    mc, _ = motion_controller
+    mc, alias = motion_controller
 
     fake_device = CAN_DEVICE.PCAN
     fake_channel = 0
@@ -251,18 +251,18 @@ def test_get_baudrate_success(motion_controller, mocker):
     fake_network = CanopenNetwork(fake_device, fake_channel, fake_baudrate)
     mocker.patch.object(mc, "_get_network", return_value=fake_network)
 
-    test_baudrate = mc.info.get_baudrate()
+    test_baudrate = mc.info.get_baudrate(alias)
 
     assert fake_baudrate == test_baudrate
 
 
 @pytest.mark.virtual
 def test_get_baudrate_failed(motion_controller, mocker):
-    mc, _ = motion_controller
+    mc, alias = motion_controller
 
     mocker.patch.object(mc, "_get_network", return_value=EthercatNetwork("fake_interface_name"))
     with pytest.raises(IMException) as imexpeption_info:
-        _ = mc.info.get_baudrate()
+        _ = mc.info.get_baudrate(alias)
 
-    expected_message_error = "The servo default is not a CANopen device."
+    expected_message_error = "The servo test is not a CANopen device."
     assert expected_message_error == imexpeption_info.value.args[0]